added PersonLink component
